Refactor: Split UserDeletionRequests component
Split the UserDeletionRequests component into smaller, more manageable files to improve maintainability.
