+++ src/components/admin/panel/users/UserDeletionRequests.tsx
@@ -1,134 +1,20 @@
 
-import { useState, useEffect } from 'react';
-import { supabase } from '@/integrations/supabase/client';
 import { Button } from '@/components/ui/button';
-import { toast } from 'sonner';
-import { Loader2, Trash2, X, RefreshCw } from 'lucide-react';
-import { Badge } from '@/components/ui/badge';
-
-type DeletionRequest = {
-  id: string;
-  user_id: string;
-  created_at: string;
-  status: 'PENDING' | 'APPROVED' | 'REJECTED';
-  reason: string;
-  username: string | null;
-  email: string | null;
-};
+import { RefreshCw } from 'lucide-react';
+import { useDeletionRequests } from './hooks/useDeletionRequests';
+import { DeletionRequestsTable } from './components/DeletionRequestsTable';
+import { EmptyRequestsState } from './components/EmptyRequestsState';
+import { LoadingState } from './components/LoadingState';
 
 export function UserDeletionRequests() {
-  const [requests, setRequests] = useState<DeletionRequest[]>([]);
-  const [loading, setLoading] = useState(true);
-  const [processing, setProcessing] = useState<string | null>(null);
-  
-  const fetchDeletionRequests = async () => {
-    setLoading(true);
-    try {
-      // Get deletion requests with user information
-      const { data, error } = await supabase
-        .from('account_deletion_requests')
-        .select(`
-          id,
-          user_id,
-          created_at,
-          status,
-          reason,
-          profiles:user_id (username, email)
-        `)
-        .order('created_at', { ascending: false });
-
-      if (error) throw error;
-      
-      // Transform data to our expected format
-      const formattedData = data.map(item => ({
-        id: item.id,
-        user_id: item.user_id,
-        created_at: item.created_at,
-        status: item.status as 'PENDING' | 'APPROVED' | 'REJECTED',
-        reason: item.reason || 'No reason provided',
-        username: item.profiles?.username,
-        email: item.profiles?.email
-      }));
-      
-      console.log('Deletion requests:', formattedData);
-      setRequests(formattedData);
-    } catch (error) {
-      console.error('Error fetching deletion requests:', error);
-      toast.error('Failed to load deletion requests');
-    } finally {
-      setLoading(false);
-    }
-  };
-  
-  useEffect(() => {
-    fetchDeletionRequests();
-  }, []);
-  
-  const approveRequest = async (requestId: string, userId: string) => {
-    if (!window.confirm('Are you sure you want to permanently delete this user account? This action cannot be undone.')) {
-      return;
-    }
-    
-    setProcessing(requestId);
-    try {
-      // Update request status to approved
-      const { error: updateError } = await supabase
-        .from('account_deletion_requests')
-        .update({ status: 'APPROVED' })
-        .eq('id', requestId);
-        
-      if (updateError) throw updateError;
-      
-      // In a real app, you would call a secure admin function to delete the user
-      // This would typically be an edge function with admin privileges
-      // For now, we'll simulate this with a success message
-      toast.success('User account deletion approved and processed');
-      
-      // Update local state
-      setRequests(prev => 
-        prev.map(r => r.id === requestId ? { ...r, status: 'APPROVED' } : r)
-      );
-    } catch (error) {
-      console.error('Error approving deletion request:', error);
-      toast.error('Failed to approve deletion request');
-    } finally {
-      setProcessing(null);
-    }
-  };
-  
-  const rejectRequest = async (requestId: string) => {
-    setProcessing(requestId);
-    try {
-      // Update request status to rejected
-      const { error } = await supabase
-        .from('account_deletion_requests')
-        .update({ status: 'REJECTED' })
-        .eq('id', requestId);
-        
-      if (error) throw error;
-      
-      toast.success('Deletion request rejected');
-      
-      // Update local state
-      setRequests(prev => 
-        prev.map(r => r.id === requestId ? { ...r, status: 'REJECTED' } : r)
-      );
-    } catch (error) {
-      console.error('Error rejecting deletion request:', error);
-      toast.error('Failed to reject deletion request');
-    } finally {
-      setProcessing(null);
-    }
-  };
-
-  const getStatusColor = (status: string) => {
-    switch (status) {
-      case 'PENDING': return 'bg-yellow-100 text-yellow-800 border-yellow-200';
-      case 'APPROVED': return 'bg-green-100 text-green-800 border-green-200';
-      case 'REJECTED': return 'bg-red-100 text-red-800 border-red-200';
-      default: return 'bg-gray-100 text-gray-800 border-gray-200';
-    }
-  };
+  const { 
+    requests, 
+    loading, 
+    processing, 
+    fetchDeletionRequests, 
+    approveRequest, 
+    rejectRequest 
+  } = useDeletionRequests();
   
   return (
     <div className="space-y-4">
@@ -140,105 +26,22 @@ export function UserDeletionRequests() {
           onClick={fetchDeletionRequests}
           disabled={loading}
         >
-          {loading ? <Loader2 className="h-4 w-4 animate-spin mr-2" /> : <RefreshCw className="h-4 w-4 mr-2" />}
+          <RefreshCw className="h-4 w-4 mr-2" />
           Refresh
         </Button>
       </div>
       
       {loading ? (
-        <div className="flex justify-center py-8">
-          <Loader2 className="h-8 w-8 animate-spin text-gray-400" />
-        </div>
+        <LoadingState />
       ) : requests.length === 0 ? (
-        <div className="text-center py-8 bg-gray-50 rounded-lg border border-gray-200">
-          <p className="text-gray-500">No account deletion requests found</p>
-        </div>
+        <EmptyRequestsState />
       ) : (
-        <div className="bg-white rounded-lg border border-gray-200 overflow-hidden">
-          <div className="overflow-x-auto">
-            <table className="min-w-full divide-y divide-gray-200">
-              <thead className="bg-gray-50">
-                <tr>
-                  <th scope="col" className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
-                    User
-                  </th>
-                  <th scope="col" className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
-                    Request Date
-                  </th>
-                  <th scope="col" className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
-                    Status
-                  </th>
-                  <th scope="col" className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
-                    Reason
-                  </th>
-                  <th scope="col" className="px-6 py-3 text-right text-xs font-medium text-gray-500 uppercase tracking-wider">
-                    Actions
-                  </th>
-                </tr>
-              </thead>
-              <tbody className="bg-white divide-y divide-gray-200">
-                {requests.map((request) => (
-                  <tr key={request.id}>
-                    <td className="px-6 py-4 whitespace-nowrap">
-                      <div className="text-sm font-medium text-gray-900">{request.username || 'Unknown user'}</div>
-                      <div className="text-xs text-gray-500">{request.email || 'No email'}</div>
-                    </td>
-                    <td className="px-6 py-4 whitespace-nowrap">
-                      <div className="text-sm text-gray-500">
-                        {new Date(request.created_at).toLocaleDateString()} 
-                        <span className="block text-xs">
-                          {new Date(request.created_at).toLocaleTimeString()}
-                        </span>
-                      </div>
-                    </td>
-                    <td className="px-6 py-4 whitespace-nowrap">
-                      <Badge variant="outline" className={`${getStatusColor(request.status)} px-2 py-1`}>
-                        {request.status}
-                      </Badge>
-                    </td>
-                    <td className="px-6 py-4">
-                      <div className="text-sm text-gray-900 max-w-xs truncate">
-                        {request.reason}
-                      </div>
-                    </td>
-                    <td className="px-6 py-4 whitespace-nowrap text-right text-sm font-medium">
-                      {request.status === 'PENDING' && (
-                        <div className="flex justify-end gap-2">
-                          <Button
-                            variant="destructive"
-                            size="sm"
-                            onClick={() => approveRequest(request.id, request.user_id)}
-                            disabled={!!processing}
-                          >
-                            {processing === request.id ? 
-                              <Loader2 className="h-4 w-4 animate-spin mr-1" /> : 
-                              <Trash2 className="h-4 w-4 mr-1" />
-                            }
-                            Delete User
-                          </Button>
-                          <Button
-                            variant="outline"
-                            size="sm"
-                            onClick={() => rejectRequest(request.id)}
-                            disabled={!!processing}
-                          >
-                            <X className="h-4 w-4 mr-1" />
-                            Reject
-                          </Button>
-                        </div>
-                      )}
-                      {request.status !== 'PENDING' && (
-                        <span className="text-gray-500">
-                          {request.status === 'APPROVED' ? 'User deleted' : 'Request rejected'}
-                        </span>
-                      )}
-                    </td>
-                  </tr>
-                ))}
-              </tbody>
-            </table>
-          </div>
-        </div>
+        <DeletionRequestsTable 
+          requests={requests}
+          processing={processing}
+          onApprove={approveRequest}
+          onReject={rejectRequest}
+        />
       )}
     </div>
   );
