+++ src/components/admin/panel/users/components/EmptyRequestsState.tsx
@@ -1,0 +1,10 @@
+
+import React from 'react';
+
+export const EmptyRequestsState = () => {
+  return (
+    <div className="text-center py-8 bg-gray-50 rounded-lg border border-gray-200">
+      <p className="text-gray-500">No account deletion requests found</p>
+    </div>
+  );
+};
